Java Variables: Mad Libs
In this project, we’ll use Java to write a Mad Libs word game! Mad Libs have short stories with blank spaces that a player can fill in. The result is usually funny (or strange).

Mad Libs require:

A short story with blank spaces (asking for different types of words).
Words from the player to fill in those blanks.

1. Create all the variables needed for the story.

2. Print the story with the variables in the right places.